refactor: extract TableNameMenuButton component for better modularity
