+++ frontend/packages/erd-core/src/components/ERDRenderer/LeftPane/LeftPane.tsx
@@ -6,28 +6,14 @@ import {
   SidebarGroupContent,
   SidebarGroupLabel,
   SidebarMenu,
-  SidebarMenuButton,
-  SidebarMenuItem,
   SidebarRail,
 } from '@liam-hq/ui'
-import { Table2 } from '@liam-hq/ui'
-import {
-  updateActiveTableName,
-  useDBStructureStore,
-  useUserEditingStore,
-} from '../../../stores'
-import styles from './LeftPane.module.css'
+import { useDBStructureStore } from '../../../stores'
 import { TableCounter } from './TableCounter'
-
-const handleClickMenuButton = (tableId: string) => () => {
-  updateActiveTableName(tableId)
-}
+import { TableNameMenuButton } from './TableNameMenuButton'
 
 export const LeftPane = () => {
   const { tables } = useDBStructureStore()
-  const {
-    active: { tableName },
-  } = useUserEditingStore()
 
   return (
     <Sidebar>
@@ -37,15 +23,7 @@ export const LeftPane = () => {
           <SidebarGroupContent>
             <SidebarMenu>
               {Object.values(tables).map((table) => (
-                <SidebarMenuItem key={table.name}>
-                  <SidebarMenuButton
-                    onClick={handleClickMenuButton(table.name)}
-                    className={table.name === tableName ? styles.active : ''}
-                  >
-                    <Table2 width="10px" />
-                    <span className={styles.tableName}>{table.name}</span>
-                  </SidebarMenuButton>
-                </SidebarMenuItem>
+                <TableNameMenuButton key={table.name} table={table} />
               ))}
             </SidebarMenu>
           </SidebarGroupContent>
