+++ frontend/packages/erd-core/src/components/ERDRenderer/LeftPane/TableNameMenuButton/index.ts
@@ -1,0 +1,1 @@
+export * from './TableNameMenuButton'
